[spring08] added comment insert

+++ src/main/java/ru/otus/spring/dao/AuthorDao.java
@@ -4,7 +4,7 @@ import ru.otus.spring.domain.Author;
 
 public interface AuthorDao {
 
-    Author getById(int i);
+    Author getById(long i);
 
     Author save(Author book);
